cleaned up more files, got rid of page 1 in house post
