--- FrontEnd/src/assets/sass/components/_PreviewSlideShow.scss
@@ -1,6 +1,0 @@
-.preview-slideshow-thumbnail {
-  @extend .disappear-on-xs;
-  min-height: 15%;
-  display: flex;
-  overflow-x: auto;
-}
