+++ FrontEnd/src/assets/sass/components/_all.scss
@@ -12,7 +12,6 @@
 @import 'ProfileModal';
 @import 'Home';
 @import 'PostForm';
-@import 'PreviewSlideShow';
 @import 'TV';
 @import 'Bookmark';
 @import 'BookmarksList';
