+++ FrontEnd/src/components/FilterForm/FilterPage1.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Container, Row, Col, Form } from 'react-bootstrap';
 import * as z from 'zod';
-import { WizardFormStep } from '../WizardForm';
+import { WizardFormStep } from '../basics/WizardForm';
 import Input from '../basics/Input';
 
 export const page1Schema = z.object({
