+++ FrontEnd/src/components/FilterForm/FilterPage3.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Container, Row, Col, Form } from 'react-bootstrap';
 import * as z from 'zod';
-import { WizardFormStep } from '../WizardForm';
+import { WizardFormStep } from '../basics/WizardForm';
 import Input from '../basics/Input';
 import { miscIcons } from '../../assets/icons/all';
