+++ FrontEnd/src/components/FilterForm/FilterPage6.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Container, Row, Form, Button } from 'react-bootstrap';
 import * as z from 'zod';
-import { WizardFormStep } from '../WizardForm';
+import { WizardFormStep } from '../basics/WizardForm';
 import ToggleGroup from '../basics/ToggleGroup';
 
 // TODO put in its own file
