+++ FrontEnd/src/components/HousingPostForm/PostPage4.tsx
@@ -1,44 +1,55 @@
 import React from 'react';
 import * as z from 'zod';
-import {
-  largeAmenitiesIcons,
-  amenitiesTranslations,
-} from '../../assets/icons/all';
-import { WizardFormStep } from '../WizardForm';
+import { WizardFormStep } from '../basics/WizardForm';
 import ToggleGroup from '../basics/ToggleGroup';
 
-// TODO this should be in a different file
-type Amenity = keyof typeof largeAmenitiesIcons;
+// TODO put in its own file
+enum Preference {
+  femaleOnly = 'Female only',
+  maleOnly = 'Male only',
+  coed = 'Co-ed',
+  clean = 'Clean',
+  quiet = 'Quiet',
+  partyOk = 'Party OK',
+  noParty = 'No party',
+  extrovert = 'Extrovert',
+  introvert = 'Introvert',
+  nightOwl = 'Night owl',
+  earlyBird = 'Early bird',
+  _420Friendly = '420 friendly',
+  smokeFree = 'Smoke free',
+  lgbtqFriendly = 'LGBTQ+ friendly',
+  overnightGuestOk = 'Overnight guest OK',
+  noOvernightGuest = 'No overnight Guest',
+}
 
 export const page4Schema = z.object({
-  amenities: z.string().array(),
+  preferences: z.string().array(),
 });
 
 export type Page4Store = z.infer<typeof page4Schema>;
 
 export const page4InitialStore: Page4Store = {
-  amenities: [],
+  preferences: [],
 };
 
 const PostPage4: React.FC<WizardFormStep<Page4Store>> = ({
-  amenities,
+  preferences,
   setStore,
 }) => {
   return (
     <ToggleGroup
-      toggleClassName="house-post-amenities-toggle"
-      label="Please select all the amenities your place offers."
-      content={(Object.keys(largeAmenitiesIcons) as [Amenity]).map((key) => ({
-        label: amenitiesTranslations[key],
-        icon: largeAmenitiesIcons[key],
-      }))}
-      initialSelected={amenities}
+      label="Please select from the following options to promote what type of person you are looking for to apply for this listing."
+      content={Object.values(Preference)}
+      initialSelected={preferences}
       onSelect={({ label, selected }) => {
         if (selected) {
-          setStore({ amenities: amenities ? [...amenities, label] : [label] });
+          setStore({
+            preferences: preferences ? [...preferences, label] : [label],
+          });
         } else {
           setStore({
-            amenities: amenities?.filter((amenity) => amenity !== label),
+            preferences: preferences?.filter((amenity) => amenity !== label),
           });
         }
       }}
