+++ FrontEnd/src/components/HousingPostForm/PostPage5.tsx
@@ -1,62 +1,59 @@
 import React from 'react';
 import * as z from 'zod';
-import {
-  largeAmenitiesIcons,
-  amenitiesTranslations,
-} from '../../assets/icons/all';
-import { WizardFormStep } from '../WizardForm';
-import ToggleGroup from '../basics/ToggleGroup';
-
-// TODO put in its own file
-enum Preference {
-  femaleOnly = 'Female only',
-  maleOnly = 'Male only',
-  coed = 'Co-ed',
-  clean = 'Clean',
-  quiet = 'Quiet',
-  partyOk = 'Party OK',
-  noParty = 'No party',
-  extrovert = 'Extrovert',
-  introvert = 'Introvert',
-  nightOwl = 'Night owl',
-  earlyBird = 'Early bird',
-  _420Friendly = '420 friendly',
-  smokeFree = 'Smoke free',
-  lgbtqFriendly = 'LGBTQ+ friendly',
-  overnightGuestOk = 'Overnight guest OK',
-  noOvernightGuest = 'No overnight Guest',
-}
+import { Container, Row, Col } from 'react-bootstrap';
+import ImageUploader from 'homehub-images-upload';
+import { WizardFormStep } from '../basics/WizardForm';
 
 export const page5Schema = z.object({
-  preferences: z.string().array(),
+  pictures: z.array(z.any()).min(1, 'You need to add some pictures!'),
 });
 
-export type Page5Store = z.infer<typeof page5Schema>;
+export interface Page5Store {
+  pictures: File[];
+}
 
 export const page5InitialStore: Page5Store = {
-  preferences: [],
+  pictures: [],
 };
 
 const PostPage5: React.FC<WizardFormStep<Page5Store>> = ({
-  preferences,
+  pictures,
+  validations,
   setStore,
+  submitForm,
 }) => {
   return (
-    <ToggleGroup
-      label="Please select from the following options to promote what type of person you are looking for to apply for this listing."
-      content={Object.values(Preference)}
-      initialSelected={preferences}
-      onSelect={({ label, selected }) => {
-        if (selected) {
-          setStore({ preferences: preferences?[...preferences, label]:[label] });
-        } else {
-          setStore({
-            preferences: preferences?.filter((amenity) => amenity !== label),
-          });
-        }
-      }}
-      center
-    />
+    <Container>
+      <Row>
+        <Col>
+          <span className="post-title">Upload photos of your place</span>
+        </Col>
+      </Row>
+
+      <Row>
+        <ImageUploader
+          withIcon
+          withPreview
+          label=""
+          onChange={(p) => {
+            setStore({ pictures: [...p] });
+          }}
+          defaultFiles={pictures}
+          defaultImages={pictures?.map((picture) =>
+            URL.createObjectURL(picture),
+          )}
+          imgExtension={['.jpg', '.png', '.jpeg']}
+          maxFileSize={5242880}
+          className="house-post-image-uploader"
+          buttonText="+"
+          buttonClassName="house-post-image-uploader-btn"
+          errorClass="house-post-error-label"
+        />
+      </Row>
+      <div className="house-post-error-label">
+        {validations?.pictures?.error}
+      </div>
+    </Container>
   );
 };
