+++ FrontEnd/src/components/NewUserSetup/Page2.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import * as z from 'zod';
 import { Container, Form, Row } from 'react-bootstrap';
-import { WizardFormStep } from '../WizardForm';
+import { WizardFormStep } from '../basics/WizardForm';
 import Input from '../basics/Input';
 
 export const page2Schema = z.object({
